atualizei o html com label, atualizei a parte do post e do delete no server.js e configurei o get e o delete no script.js

+++ script.js
@@ -1,34 +1,6 @@
-// GET
-
-const container1 = document.getElementById("lista"); //const que pega o elemento do HTML pelo ID
-
-async function mostrarUsuarios() { //função que faz o GET na API
-  try { // o try "tenta" executar oque foi pedido
-    const resposta = await fetch("http://localhost:3000/users"); // o fetch é usado para localizar o endereço da API em questão
-    const dados = await resposta.json(); // transforma em objeto JS
-    console.log(dados);
-
-    // limpa os containers antes de preencher
-    container1.innerHTML = "";
-
-    // percorre todos os objetos do array
-    for (let i = 0; i < dados.length; i++) {
-      const post = document.createElement("h1");
-      //post.classList.add("usuario");
-      post.textContent = `Id: ${dados[i].id}, Nome: ${dados[i].nome}, CPF: ${dados[i].cpf}, Telefone: ${dados[i].telefone}.`;
-
-      container1.appendChild(post);
-    }
-
-  } catch (erro) {
-    console.error("Erro ao carregar formulário:", erro);
-  }
-}
-
-
 // ---------------------------------  POST  ---------------------------------------
 
-const bntPOST = document.getElementById("enviarPOST");
+const bntPOST = document.getElementById("enviarPOST"); // botão, pega elemento no documento html
 
 async function adicionaUsuario() {
   try {
@@ -37,51 +9,77 @@ async function adicionaUsuario() {
     const cpf = document.getElementById("novoCpf").value;
   
 
-    const resposta = await fetch("http://localhost:3000/users", {
+    const resposta = await fetch("http://localhost:3000/users", { // o fetch busca na api
       method: "POST", // aqui informamos o metodo
       headers: {
         "Content-Type": "application/json",
       },
       body: JSON.stringify({nome, telefone, cpf}), // aqui temos o corpo do objeto
     });
-    const dados = await resposta.json();
-    console.log("Usuário criado", dados);
+    const dados = await resposta.json(); // resposta/promise
 
-    mostrarUsuarios();
-
-  } catch (erro){
-    console.error("Erro ao criar usuário", erro);
-  }
-}
-
-bntPOST.addEventListener("click", adicionaUsuario);
-
-
-
-// ----------------- DELETE -----------------------------------------------
-
-const idDelete = document.getElementById("idDelete");
-const bntDel = document.getElementById("btnDELETE");
-
-async function deletarPost() {
-  try {
-    const id = idDelete.value.trim();
-
-    const resposta = await fetch(`http://localhost:3000/users/${id}`, {
-      method: "DELETE"
-    });
-
-    if (resposta.ok){
-      console.log("Post deletado com sucesso");
-
-    } else {
-      console.log("Falha ao deletor o post");
+    if (dados.erro){
+      throw new Error(dados.erro);
     }
 
-    getUsuarios();
-  } catch (erro) {
-    console.log("Erro ao deletar post", erro);
+    console.log("Usuário cadastrado com sucesso!", dados);
+
+    mostrarUsuarios(); // roda a função get para mostrar e atualizar a lista
+
+  } catch (erro){
+    console.error("Erro ao enviar usuário!", erro);
+    window.alert(erro); // abre um alerta pro usuario com o erro da back
   }
 }
 
-bntDel.addEventListener("click", deletarPost);
+bntPOST.addEventListener("click", adicionaUsuario); // roda a fução atraves do click
+
+// ----------------------------------- GET -------------------------------------------
+
+const container = document.getElementById("lista"); //const que pega o elemento do HTML pelo ID
+
+async function mostrarUsuarios() { //função que faz o GET na API
+  try { // o try "tenta" executar oque foi pedido
+    const resposta = await fetch("http://localhost:3000/users"); // o fetch é usado para localizar o endereço da API em questão
+    const dados = await resposta.json(); // transforma em objeto JS
+    console.log(dados);
+
+    container.innerHTML = "";  // limpa os containers antes de preencher
+
+    for (let i = 0; i < dados.length; i++) {  // percorre todos os objetos do array
+      const post = document.createElement("h1"); // cria um h1 no html
+      post.classList.add("usuario"); // adiciona a class "usuario" ao elemento criado
+      post.textContent = `Id: ${dados[i].id}, Nome: ${dados[i].nome}, CPF: ${dados[i].cpf}, Telefone: ${dados[i].telefone}.`;
+
+      const btnDelete = document.createElement("button"); // cria o botão "delete"
+      btnDelete.textContent = "X";
+      btnDelete.addEventListener("click", () => deletarUsuario(dados[i].id)); // roda a função no click
+
+      post.appendChild(btnDelete); // o appendChild insere os elementos
+      container.appendChild(post);
+    }
+
+  } catch (erro) { // pega o erro
+    console.error("Erro ao carregar usuário!", erro);
+    window.alert(erro);
+  }
+}
+
+// -------------------------------- DELETE --------------------------------
+
+async function deletarUsuario(id){
+  try{
+      const confirmar = confirm("Tem certeza que deseja excluir este usuário?"); // abre caixa de confirmação
+        if (!confirmar) return;
+
+        const resposta = await fetch(`http://localhost:3000/users/${id}`, { //localiza o fetch pelo id
+          method: "DELETE",
+      });
+
+  mostrarUsuarios(); // chama a função do get para atualizar lista
+
+  } catch (erro) {
+    console.error("Erro ao excluir usuário!", erro);
+    window.alert(erro);
+  }
+}
